+++ src/main/java/com/meohin/todo_pe/repository/TaskMeasuresRepository.java
@@ -3,5 +3,5 @@ package com.meohin.todo_pe.repository;
 import com.meohin.todo_pe.entity.TaskMeasures;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskMeasuresRepository extends JpaRepository<TaskMeasures, Integer> {
+public interface TaskMeasuresRepository extends JpaRepository<TaskMeasures, Long> {
 }
